pyproject.toml
[project]
name = "qingyu-ai-service"
version = "0.1.0"
description = "Qingyu AI Service - Phase3 v2.0 Implementation"
authors = [
    {name = "Qingyu Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # FastAPI and Server
    "fastapi>=0.109.0,<0.110.0",
    "uvicorn[standard]>=0.27.0,<0.28.0",
    "pydantic>=2.5.3,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    # gRPC
    "grpcio>=1.60.0,<2.0.0",
    "grpcio-tools>=1.60.0,<2.0.0",
    "protobuf>=4.25.1,<5.0.0",
    # LangChain and LangGraph - Latest versions
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "langchain-community",
    "langgraph",
    "langgraph-checkpoint",
    "langgraph-checkpoint-postgres",
    "langsmith",
    # Vector Database
    "pymilvus>=2.3.4,<3.0.0",
    # Embeddings
    "sentence-transformers>=2.3.1,<3.0.0",
    "torch>=2.1.2,<3.0.0",
    # Redis for Event Bus
    "redis>=5.0.1,<6.0.0",
    # Logging
    "structlog>=24.1.0,<25.0.0",
    # Testing
    "pytest>=7.4.4,<8.0.0",
    "pytest-asyncio>=0.23.3,<1.0.0",
    "httpx>=0.26.0,<1.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.1,<24.0.0",
    "isort>=5.13.2,<6.0.0",
    "flake8>=7.0.0,<8.0.0",
    "mypy>=1.8.0,<2.0.0",
    "ipython>=8.19.0,<9.0.0",
    # Performance testing
    "pytest-benchmark>=4.0.0,<5.0.0",
    "memory-profiler>=0.61.0,<1.0.0",
    "psutil>=5.9.0,<6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
